Updated all dependencies
