+++ package.json
@@ -4,24 +4,15 @@
   "private": true,
   "dependencies": {
     "@apollo/client": "^3.7.3",
-    "@cedalio/sdk-js": "^0.1.0-beta.3",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
+    "@cedalio/sdk-js": "^0.1.0-beta.6",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.11.0",
     "@mui/material": "^5.11.1",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.10",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.9",
-    "@web3modal/ethereum": "^2.0.0-beta.10",
-    "@web3modal/react": "^2.0.0-beta.10",
+    "@web3modal/ethereum": "^2.7.1",
+    "@web3modal/react": "^2.7.1",
     "apollo-upload-client": "^17.0.0",
-    "axios": "^1.3.2",
     "classnames": "^2.3.2",
-    "ethers": "^5.7.2",
     "graphql": "^16.6.0",
     "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.1",
@@ -31,10 +22,14 @@
     "react-hook-form": "^7.45.4",
     "react-loader-spinner": "^5.3.4",
     "react-scripts": "^5.0.1",
-    "typescript": "^4.9.4",
-    "wagmi": "^0.9.5",
-    "web-vitals": "^2.1.4",
-    "web3": "^4.1.1"
+    "typescript": "^5.1.6",
+    "wagmi": "^1.4.2",
+    "web-vitals": "^2.1.4"
+  },
+  "overrides": {
+    "react-scripts": {
+      "typescript": "^5"
+    }
   },
   "scripts": {
     "start": "react-scripts start",
@@ -44,7 +39,8 @@
     "check-types": "tsc --noemit",
     "lint": "eslint --ext .js,.ts,.tsx . && npm run check-types",
     "lint-fix": "eslint --ext .js,.ts,.tsx . --fix && npm run check-types",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "analyze": "source-map-explorer 'build/static/js/*.js'"
   },
   "eslintConfig": {
     "extends": [
@@ -67,8 +63,15 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "@types/apollo-upload-client": "^17.0.2",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.10",
+    "@types/react": "^18.0.26",
     "@types/react-beautiful-dnd": "^13.1.3",
+    "@types/react-dom": "^18.0.9",
     "@types/react-gtm-module": "^2.0.1",
     "@types/styled-components": "^5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
@@ -80,6 +83,8 @@
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-react": "^7.33.1",
-    "husky": "^8.0.0"
+    "husky": "^8.0.0",
+    "prettier": "^3.0.3",
+    "source-map-explorer": "^2.5.3"
   }
 }
